16702 Change height from BigInteger to Long

+++ src/main/java/com/softjourn/eris/filter/type/FilterHeight.java
@@ -3,11 +3,9 @@ package com.softjourn.eris.filter.type;
 import com.softjourn.eris.filter.FilterData;
 import com.softjourn.eris.filter.Operation;
 
-import java.math.BigInteger;
-
 
 public class FilterHeight extends FilterData {
-    public FilterHeight(Operation op, BigInteger value) {
+    public FilterHeight(Operation op, Long value) {
         super("height", op, value.toString());
     }
 }
