+++ src/main/java/com/softjourn/eris/transaction/type/Height.java
@@ -3,8 +3,6 @@ package com.softjourn.eris.transaction.type;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigInteger;
-
 /**
  * Height returns from get latest block
  * Created by vromanchuk on 17.01.17.
@@ -12,5 +10,5 @@ import java.math.BigInteger;
 @Data
 @AllArgsConstructor
 public class Height {
-    BigInteger height;
+    Long height;
 }
